requirements.txt
click>=8.0
python-docx>=1.0
openai>=1.0
python-dotenv>=1.0
